backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pdfminer.six==20231228
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
simplekml==1.3.6
shapely==2.0.6
